Move :: move exception package to global
